% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_clustering.R
\name{dbscan_clustering}
\alias{dbscan_clustering}
\title{Draw the compressed network}
\usage{
dbscan_clustering(
  Complete_network,
  eps = 0.8,
  opacity = 0.5,
  size = 25,
  fontSize = 0.7
)
}
\arguments{
\item{Complete_network}{A igraph object with the information}

\item{eps}{Is the epsilon value, that idicate the max distance between two nodes for put both in the same cluster (default: 0.8)}

\item{opacity}{For the network inks and nodes (default : 0.5)}

\item{size}{Is the size of nodes (default: 25)}

\item{fontSize}{Is the font size of node labels (default : 0.7)}
}
\value{
The network in compressed form based on the DBSCAN clustering
}
\description{
This function takes an igraph object and compresses the network based on clustering method, in particular
this method is DBSCAN clustering
}
\examples{

# Example usage:
# Create an example dataset
data_example <- data.frame(
  nodeA = c("A", "B", "C", "D", "E", "F"),
  nodeB = c("B", "C", "A","A", "B", "E"),
  score = c(0.8, 0.9, 0.85, 0.6, 0.5, 0.75)
)
# Draw the complete network
complete_net = draw_complete_net(data_example, nodeA = "nodeA", nodeB = "nodeB", score = "score")
# Use the compression function recalling the dbscan method
compression = GraphCompression(complete_net$Complete_network, method = "dbscan", eps = 1.2)
# to watch all the results obtained
print(compression)
}
