% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_complete_net.R
\name{draw_complete_net}
\alias{draw_complete_net}
\title{Draw the complete network}
\usage{
draw_complete_net(
  dataframe_interaction,
  nodeA,
  nodeB,
  score,
  fontSize = 0.7,
  opacity = 0.5,
  size = 25
)
}
\arguments{
\item{dataframe_interaction}{A dataframe with the necessary information}

\item{nodeA}{The first column of dataframe referring to the nodes that are inside the network (interact with nodeB)}

\item{nodeB}{The second column of dataframe referring to the nodes that are inside the network (interact with nodeA)}

\item{score}{The score column of dataframe that defines the links between the nodes}

\item{fontSize}{Is the font size of node labels (default : 0.7)}

\item{opacity}{For the network inks and nodes (default : 0.5)}

\item{size}{Is the size of nodes (default: 25)}
}
\value{
The complete network visualization
}
\description{
This function takes a dataframe that describe the interaction between protein and with these information plot the network
}
\examples{

# Example usage:
# Create an example dataset
data_example <- data.frame(
  nodeA = c("A", "B", "C"),
  nodeB = c("B", "C", "A"),
  score = c(0.8, 0.9, 0.85)
)
# Draw the complete network
draw_complete_net(data_example, nodeA = "nodeA", nodeB = "nodeB", score = "score")
}
