% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partners.R
\name{get_partners}
\alias{get_partners}
\title{Get the participant of the network}
\usage{
get_partners(protein_name, species = 9606)
}
\arguments{
\item{protein_name}{A character string representing the protein name.}

\item{species}{The numeric ID for the species, default is 9606 (Human).}
}
\value{
A data frame describing the network.
}
\description{
This function takes a protein name and retrieves network information from the STRING database via API
}
\examples{

# Example usage:
# Retrieve network information for the protein TP53 in humans
network_data <- get_partners("TP53")
print(network_data)

#' # Retrieve network information for a different species
network_data_mouse <- get_partners("TP53", species = 10090)
print(network_data_mouse)
}
