% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphCompression.R
\name{GraphCompression}
\alias{GraphCompression}
\title{Wrapper function for compression}
\usage{
GraphCompression(Network, links = NULL, method = NULL, ...)
}
\arguments{
\item{Network}{the igraph object}

\item{links}{the information related to the network (necessary only for some functions)}

\item{method}{the type of compression}

\item{...}{other args}
}
\value{
The network in compressed form with a specific method
}
\description{
This function takes an igraph object and the related information
}
\examples{

# Example usage:
# Create an example dataset
data_example <- data.frame(
  nodeA = c("A", "B", "C", "D", "E", "F"),
  nodeB = c("B", "C", "A","A", "B", "E"),
  score = c(0.8, 0.9, 0.85, 0.6, 0.5, 0.75)
)
# Draw the complete network
complete_net = draw_complete_net(data_example, nodeA = "nodeA", nodeB = "nodeB", score = "score")

# Use the function recalling a compression method.
# REMEMBER: This is a wrapper function, that can recall all the compression functions as methods
# in this first example using the betweenness method
compression = GraphCompression(complete_net$Complete_network,complete_net$edges_data, method = "betweenness")
# to watch all the results obtained
print(compression)
# in this second example using the kmeans method
compression = GraphCompression(complete_net$Complete_network, method = "kmeans")
}
