% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_clustering.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{Draw the compressed network}
\usage{
kmeans_clustering(
  Complete_network,
  num_clusters = 3,
  size = 25,
  fontSize = 0.7,
  opacity = 0.5
)
}
\arguments{
\item{Complete_network}{A igraph object with the information}

\item{num_clusters}{Is the number of the clusters, so the nodes in the compressed graph (default: 3)}

\item{size}{Is the size of nodes (default: 25)}

\item{fontSize}{Is the font size of node labels (default : 0.7)}

\item{opacity}{For the network inks and nodes (default : 0.5)}
}
\value{
The network in compressed form based on the K-means clustering
}
\description{
This function takes an igraph object and compresses the network based on clustering method, in particular
this method is K-means clustering
}
\examples{

# Example usage:
# Create an example dataset
data_example <- data.frame(
  nodeA = c("A", "B", "C", "D", "E", "F"),
  nodeB = c("B", "C", "A","A", "B", "E"),
  score = c(0.8, 0.9, 0.85, 0.6, 0.5, 0.75)
)
# Draw the complete network
complete_net = draw_complete_net(data_example, nodeA = "nodeA", nodeB = "nodeB", score = "score")
# Use the compression function recalling the kmeans method
compression = GraphCompression(complete_net$Complete_network, method = "kmeans")
# to watch all the results obtained
print(compression)
}
