% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Netty.R
\name{Netty}
\alias{Netty}
\title{This is a data set example, useful to practice with the package's function}
\usage{
Netty()
}
\value{
A data frame (representing a network) that correspond to the data set example.
}
\description{
This function takes import a data set as an example
}
\examples{

# Example usage of Netty function
network_data <- Netty()
print(network_data)

}
