% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageRank_centralitry_compression.R
\name{pageRank_centrality_compression}
\alias{pageRank_centrality_compression}
\title{Draw the compressed network}
\usage{
pageRank_centrality_compression(Complete_network, links, top = 0.1, size = 25)
}
\arguments{
\item{Complete_network}{A igraph object with the information}

\item{links}{Is the data frame with the nodes interaction information}

\item{top}{A percentage of the top nodes (default: 0.1) --> so take the top 90\% nodes}

\item{size}{Is the size of nodes (default: 25)}
}
\value{
The network in compressed form based on pageRank centrality
}
\description{
This function takes an igraph object and compresses the network based on the pageRank centrality
}
\examples{

# Example usage:
# Create an example dataset
data_example <- data.frame(
  nodeA = c("A", "B", "C", "D", "E", "F"),
  nodeB = c("B", "C", "A","A", "B", "E"),
  score = c(0.8, 0.9, 0.85, 0.6, 0.5, 0.75)
)
# Draw the complete network
complete_net = draw_complete_net(data_example, nodeA = "nodeA", nodeB = "nodeB", score = "score")
# Use the compression function recalling the pagerank method
compression = GraphCompression(complete_net$Complete_network,complete_net$edges_data, method = "pagerank")
# to watch all the results obtained
print(compression)
}
